refactor: arrange kt files
