+++ app/src/main/AndroidManifest.xml
@@ -59,11 +59,11 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".presentation.view.MainActivity"
+            android:name=".presentation.view.main.KakaoLoginActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".presentation.view.KakaoLoginActivity"
+            android:name=".presentation.view.main.MainActivity"
             android:exported="true"
             android:screenOrientation="portrait">
             <intent-filter>
