+++ app/src/main/java/com/example/everymoment/presentation/adapter/FriendRequestAdapter.kt
@@ -4,13 +4,17 @@ import android.app.AlertDialog
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentActivity
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.example.everymoment.data.repository.Member
 import com.example.everymoment.databinding.FriendRequestItemBinding
+import com.example.everymoment.extensions.CustomDialog
 
 class FriendRequestAdapter(
+    private val activity: FragmentActivity,
     private val onFriendRequest: (Member) -> Unit
 ) : ListAdapter<Member, FriendRequestAdapter.FriendRequestViewHolder>(
     FriendRequestDiffCallback()
@@ -19,7 +23,7 @@ class FriendRequestAdapter(
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): FriendRequestViewHolder {
         val binding =
             FriendRequestItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
-        return FriendRequestViewHolder(binding, onFriendRequest)
+        return FriendRequestViewHolder(binding, activity, onFriendRequest)
     }
 
     override fun onBindViewHolder(holder: FriendRequestViewHolder, position: Int) {
@@ -28,6 +32,7 @@ class FriendRequestAdapter(
 
     class FriendRequestViewHolder(
         private val binding: FriendRequestItemBinding,
+        private val activity: FragmentActivity,
         private val onFriendRequest: (Member) -> Unit
     ) : RecyclerView.ViewHolder(binding.root) {
         fun bind(user: Member) {
@@ -39,16 +44,11 @@ class FriendRequestAdapter(
         }
 
         private fun showFriendRequestConfirmationDialog(user: Member) {
-            AlertDialog.Builder(itemView.context)
-                .setTitle("친구 신청")
-                .setMessage("${user.nickname}님에게 친구 신청을 하시겠습니까?")
-                .setNegativeButton("아니오") { dialog, _ -> dialog.dismiss() }
-                .setPositiveButton("네") { _, _ ->
-                    onFriendRequest(user)
-                    binding.friendRequestButton.visibility = View.GONE
-                    binding.requestCompletedButton.visibility = View.VISIBLE
-                }
-                .show()
+            CustomDialog("${user.nickname}님에게 친구 신청을 하시겠습니까?", "아니오", "네", onPositiveClick = {
+                onFriendRequest(user)
+                binding.friendRequestButton.visibility = View.GONE
+                binding.requestCompletedButton.visibility = View.VISIBLE
+            }).show(activity.supportFragmentManager, "CustomDialog")
         }
     }
