+++ app/src/main/java/com/example/everymoment/presentation/view/main/MainActivity.kt
@@ -1,15 +1,13 @@
-package com.example.everymoment.presentation.view
+package com.example.everymoment.presentation.view.main
 
-import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import android.view.View
-import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.example.everymoment.R
 import com.example.everymoment.databinding.ActivityMainBinding
+import com.example.everymoment.presentation.view.main.search.SearchFragment
 import com.google.firebase.ktx.Firebase
-import com.google.firebase.messaging.FirebaseMessaging
 import com.google.firebase.remoteconfig.FirebaseRemoteConfig
 import com.google.firebase.remoteconfig.ktx.remoteConfig
 import com.google.firebase.remoteconfig.ktx.remoteConfigSettings
