+++ app/src/main/java/com/example/everymoment/presentation/view/main/search/SearchFilterDialogFragment.kt
@@ -1,7 +1,6 @@
-package com.example.everymoment.presentation.view
+package com.example.everymoment.presentation.view.main.search
 
 import android.app.DatePickerDialog
-import android.graphics.Color
 import android.icu.util.Calendar
 import android.os.Build
 import android.os.Bundle
