+++ app/src/main/java/com/example/everymoment/presentation/view/main/search/SearchFragment.kt
@@ -1,11 +1,10 @@
-package com.example.everymoment.presentation.view
+package com.example.everymoment.presentation.view.main.search
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.everymoment.presentation.view.SearchFilterDialogFragment
 import com.example.everymoment.databinding.FragmentSearchBinding
 
 class SearchFragment : Fragment() {
