+++ app/src/main/java/com/example/everymoment/presentation/view/sub/diary/DiaryReadFragment.kt
@@ -1,4 +1,4 @@
-package com.example.everymoment.presentation.view
+package com.example.everymoment.presentation.view.sub.diary
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
