+++ app/src/main/res/layout/activity_kakao_login.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".presentation.view.KakaoLoginActivity">
+    tools:context=".presentation.view.main.KakaoLoginActivity">
 
     <TextView
         android:id="@+id/appName"
