+++ app/src/main/res/layout/fragment_calendar_view.xml
@@ -6,19 +6,19 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".presentation.view.CalendarViewFragment">
-
+    tools:context=".presentation.view.main.CalendarViewFragment">
 
 
     <TextView
         android:id="@+id/findFriendByNickname"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="20dp"
+        android:layout_marginStart="16dp"
+        android:layout_marginTop="16dp"
         android:text="캘린더"
         android:textColor="@color/black"
         android:textSize="18sp"
-        app:layout_constraintEnd_toEndOf="parent"
+        android:textStyle="bold"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
@@ -31,16 +31,32 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/findFriendByNickname" />
 
-    <androidx.appcompat.widget.AppCompatImageButton
+    <LinearLayout
         android:id="@+id/calendarAlarm"
-        android:layout_width="48dp"
-        android:layout_height="48dp"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentEnd="true"
         android:layout_marginTop="10dp"
-        android:layout_marginEnd="8dp"
-        android:background="@null"
-        android:src="@drawable/notifications_40px"
+        android:layout_marginEnd="16dp"
+        android:gravity="center_horizontal"
+        android:orientation="vertical"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent">
+
+        <ImageView
+            android:id="@+id/notificationIcon"
+            android:layout_width="27dp"
+            android:layout_height="27dp"
+            android:src="@drawable/ic_notification_unread" />
+
+        <TextView
+            android:id="@+id/notificationText"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="알림"
+            android:textColor="#FF5f6368"
+            android:textSize="10sp" />
+    </LinearLayout>
 
 
     <com.prolificinteractive.materialcalendarview.MaterialCalendarView
@@ -49,10 +65,10 @@
         android:layout_height="570dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/line"
         app:mcv_headerTextAppearance="@style/CalendarHeaderTextAppearance"
         app:mcv_selectionColor="@color/black"
         app:mcv_showOtherDates="all"
-        app:layout_constraintTop_toBottomOf="@id/line"
         app:mcv_tileHeight="80dp"
         app:mcv_tileSize="53dp" />
