+++ app/src/main/res/layout/fragment_diary_edit.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="#515151"
     android:orientation="vertical"
-    tools:context=".presentation.view.DiaryEditFragment">
+    tools:context=".presentation.view.sub.diary.DiaryEditFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
