+++ app/src/main/res/layout/fragment_friends_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".presentation.view.FriendsListFragment">
+    tools:context=".presentation.view.sub.friends.FriendsListFragment">
 
     <TextView
         android:id="@+id/friendsList"
