+++ app/src/main/res/layout/fragment_notification.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     android:orientation="vertical"
-    tools:context=".presentation.view.NotificationFragment">
+    tools:context=".presentation.view.sub.NotificationFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
