+++ app/src/main/res/layout/fragment_search.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     android:orientation="vertical"
-    tools:context=".presentation.view.SearchFragment">
+    tools:context=".presentation.view.main.search.SearchFragment">
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
