+++ app/src/main/res/layout/fragment_setting.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".presentation.view.SettingFragment">
+    tools:context=".presentation.view.main.SettingFragment">
 
     <TextView
         android:id="@+id/account"
