+++ app/src/main/res/layout/fragment_today_log.xml
@@ -7,24 +7,35 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@color/white"
-    tools:context=".presentation.view.TodayLogFragment">
+    tools:context=".presentation.view.main.TodayLogFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:padding="16dp">
+        android:layout_height="wrap_content">
 
         <TextView
             android:id="@+id/title"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_marginStart="16dp"
+            android:layout_marginTop="16dp"
             android:text="오늘의 일기"
             android:textColor="@color/black"
             android:textSize="18sp"
             android:textStyle="bold"
             android:layout_alignParentStart="true"/>
 
+        <View
+            android:id="@+id/line"
+            android:layout_width="match_parent"
+            android:layout_height="2dp"
+            android:layout_marginTop="15dp"
+            android:background="@color/line_color"
+            android:layout_below="@+id/title" />
+
         <LinearLayout
+            android:layout_marginTop="10dp"
+            android:layout_marginEnd="16dp"
             android:id="@+id/notification"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -46,7 +57,6 @@
                 android:textSize="10sp"
                 android:textColor="#FF5f6368" />
         </LinearLayout>
-
     </RelativeLayout>
 
     <LinearLayout
